// src/app/shared/interfaces/myfile.interface.ts
export interface filedataRes {
  nBundledetailid?: string;
  cPath?: string;
  cPage?: string;
  cRefpage?: string;
  cFiletype?: string;
}


export interface filedataReq {
  nBundledetailid?: string;
  cTab?: string;
  nCaseid?: string;
}